Extracted helper function to generate a Csv file
