+++ test/servers/services/ServersExporter.test.ts
@@ -11,10 +11,7 @@ describe('ServersExporter', () => {
   });
   const appendChild = jest.fn();
   const removeChild = jest.fn();
-  const createWindowMock = (isIe10 = true) => Mock.of<Window>({
-    navigator: {
-      msSaveBlob: isIe10 ? jest.fn() : undefined,
-    },
+  const createWindowMock = () => Mock.of<Window>({
     document: {
       createElement: jest.fn(() => createLinkMock()),
       body: { appendChild, removeChild },
@@ -66,20 +63,8 @@ describe('ServersExporter', () => {
       expect(erroneousToCsv).toHaveBeenCalledTimes(1);
     });
 
-    it('makes use of msSaveBlob API when available', () => {
+    it('makes use of download link API', () => {
       const windowMock = createWindowMock();
-      const exporter = new ServersExporter(storageMock, windowMock, createCsvjsonMock());
-      const { navigator: { msSaveBlob }, document: { createElement } } = windowMock;
-
-      exporter.exportServers();
-
-      expect(storageMock.get).toHaveBeenCalledTimes(1);
-      expect(msSaveBlob).toHaveBeenCalledTimes(1);
-      expect(createElement).not.toHaveBeenCalled();
-    });
-
-    it('makes use of download link API when available', () => {
-      const windowMock = createWindowMock(false);
       const exporter = new ServersExporter(storageMock, windowMock, createCsvjsonMock());
       const { document: { createElement } } = windowMock;
